Move multiple tests to a single file
