--- tests/unit/Context/ctx_html_bypass.spec.js
@@ -1,16 +1,0 @@
-import { mount } from "@vue/test-utils";
-import CldContext from "../../../src/components/CldContext/CldContext.vue";
-
-describe("CldContext", () => {
-  it("bypasses non-cloudinary attributes", () => {
-    const wrapper = mount({
-      template: `
-        <cld-context cloudName="demo" aria-hidden="true" />
-      `,
-      components: { CldContext }
-    });
-    expect(wrapper.findAll("div").length).toBe(1);
-    expect(wrapper.find("div").attributes("aria-hidden")).toBe("true");
-    expect(wrapper.find("div").attributes("cloudName")).toBe(undefined);
-  });
-});
